pyproject.toml
[project]
name = "infosec-labs"
version = "0.1.0"
description = ""
authors = [
    {name = "Ilya Tampio",email = "quakumei@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "keyboard (>=0.13.5,<0.14.0)",
    "click (>=8.1.8,<9.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pycryptodome (>=3.22.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
